Fix warnings.

+++ OS/Kernel/FileSystems/FAT/fat.c
@@ -163,7 +163,7 @@ bool fat_read_file(char *path, uint8_t *buffer, uint32_t start_index, uint32_t l
 
     uint16_t initial_sector = start_index / 512;
     uint16_t last_sector = (start_index + length) / 512;
-    uint16_t sectors_count = length == 0 ? file_info->size / 512 + 1 : last_sector - initial_sector + 1;
+    uint16_t sectors_count = length == 0 ? file_info->size / 512 + 1 : (uint16_t)(last_sector - initial_sector + 1);
 
     uint8_t *result = fat_load_file_from_sector(file_info->first_sector, initial_sector, sectors_count);
     memcpy(buffer, result + (start_index % 512), length);
